feat: 相同的逻辑值保留一份，全部提到useCascader里
