+++ src/cascader/src/Cascader.tsx
@@ -1,3 +1,4 @@
+import type { CheckStrategy } from 'treemate'
 import type { InternalSelectionInst } from '../../_internal'
 import type { RenderTag } from '../../_internal/selection/src/interface'
 import type { ThemeProps } from '../../_mixins'
@@ -12,14 +13,12 @@ import type {
   CascaderOption,
   ExpandTrigger,
   Filter,
-  Key,
   OnLoad,
   OnUpdateValue,
   SelectMenuInstance,
   Value
 } from './interface'
 import { changeColor, depx, getPreciseEventTarget, happensIn } from 'seemly'
-import { type CheckStrategy, SubtreeNotLoadedError } from 'treemate'
 import { useIsMounted, useMergedState } from 'vooks'
 import {
   computed,
@@ -46,26 +45,17 @@ import {
   VTarget
 } from 'vueuc'
 import { NInternalSelection } from '../../_internal'
-import {
-  useConfig,
-  useFormItem,
-  useLocale,
-  useTheme,
-  useThemeClass
-} from '../../_mixins'
+import { useTheme, useThemeClass } from '../../_mixins'
 import {
   call,
   markEventEffectPerformed,
   useAdjustedTo,
   warnOnce
 } from '../../_utils'
-import { cascaderLight } from '../styles'
 import CascaderMenu from './CascaderMenu'
 import CascaderSelectMenu from './CascaderSelectMenu'
 import { useCascader } from './hooks/useCascader'
 import { cascaderInjectionKey } from './interface'
-import style from './styles/index.cssr'
-import { getRawNodePath } from './utils'
 
 export const cascaderProps = {
   ...(useTheme.props as ThemeProps<CascaderTheme>),
@@ -219,25 +209,17 @@ export default defineComponent({
         }
       })
     }
+
+    const patternRef = ref('')
     const {
       mergedBorderedRef,
       mergedClsPrefixRef,
       namespaceRef,
-      inlineThemeDisabled
-    } = useConfig(props)
-    const themeRef = useTheme(
-      'Cascader',
-      '-cascader',
-      style,
-      cascaderLight,
-      props,
-      mergedClsPrefixRef
-    )
-
-    const { localeRef } = useLocale('Cascader')
-    const patternRef = ref('')
-    const formItem = useFormItem(props)
-    const {
+      inlineThemeDisabled,
+      themeRef,
+      formItem,
+      localeRef,
+      optionHeightRef,
       mergedValueRef,
       mergedCheckStrategyRef,
       keyboardKeyRef,
@@ -246,7 +228,6 @@ export default defineComponent({
       addLoadingKey,
       deleteLoadingKey,
       treeMateRef,
-      mergedKeysRef,
       checkedKeysRef,
       indeterminateKeysRef,
       menuModelRef,
@@ -254,11 +235,16 @@ export default defineComponent({
       updateKeyboardKey,
       updateHoverKey,
       getOptionsByKeys,
+      showCheckboxRef,
       selectedOptionsRef,
       selectedOptionRef,
       doUpdateValue,
-      doUncheck
-    } = useCascader(props, formItem)
+      doUncheck,
+      doCheck,
+      move,
+      cssVarsRef,
+      themeClassHandle
+    } = useCascader(props)
 
     const { mergedSizeRef, mergedDisabledRef, mergedStatusRef } = formItem
     const cascaderMenuInstRef = ref<CascaderMenuInstance | null>(null)
@@ -270,10 +256,6 @@ export default defineComponent({
     const adjustedToRef = useAdjustedTo(props)
     const focusedRef = ref(false)
 
-    const optionHeightRef = computed(() => {
-      return themeRef.value.self.optionHeight
-    })
-
     const uncontrolledShowRef = ref(false)
     function doUpdateShow(value: boolean): void {
       const { onUpdateShow, 'onUpdate:show': _onUpdateShow } = props
@@ -284,72 +266,6 @@ export default defineComponent({
         call(_onUpdateShow, value)
       }
       uncontrolledShowRef.value = value
-    }
-
-    function doCheck(key: Key): boolean {
-      const { cascade, multiple, filterable } = props
-      const {
-        value: { check, getNode, getPath }
-      } = treeMateRef
-      if (multiple) {
-        try {
-          const { checkedKeys } = check(key, mergedKeysRef.value.checkedKeys, {
-            cascade,
-            checkStrategy: mergedCheckStrategyRef.value,
-            allowNotLoaded: props.allowCheckingNotLoaded
-          })
-          doUpdateValue(
-            checkedKeys,
-            getOptionsByKeys(checkedKeys),
-            checkedKeys.map(checkedKey =>
-              getRawNodePath(getPath(checkedKey)?.treeNodePath)
-            )
-          )
-          if (filterable)
-            focusSelectionInput()
-          keyboardKeyRef.value = key
-          hoverKeyRef.value = key
-        }
-        catch (err) {
-          if (err instanceof SubtreeNotLoadedError) {
-            if (cascaderMenuInstRef.value) {
-              const tmNode = getNode(key)
-              if (tmNode !== null) {
-                cascaderMenuInstRef.value.showErrorMessage(
-                  (tmNode.rawNode as any)[props.labelField] as string
-                )
-              }
-            }
-          }
-          else {
-            throw err
-          }
-        }
-      }
-      else {
-        if (mergedCheckStrategyRef.value === 'child') {
-          const tmNode = getNode(key)
-          if (tmNode?.isLeaf) {
-            doUpdateValue(
-              key,
-              tmNode.rawNode,
-              getRawNodePath(getPath(key).treeNodePath)
-            )
-          }
-          else {
-            return false
-          }
-        }
-        else {
-          const tmNode = getNode(key)
-          doUpdateValue(
-            key,
-            tmNode?.rawNode || null,
-            getRawNodePath(getPath(key)?.treeNodePath)
-          )
-        }
-      }
-      return true
     }
 
     const controlledShowRef = toRef(props, 'show')
@@ -461,92 +377,6 @@ export default defineComponent({
         patternRef.value = ''
     }
     // --- keyboard
-    function move(direction: 'prev' | 'next' | 'child' | 'parent'): void {
-      const { value: keyboardKey } = keyboardKeyRef
-      const { value: treeMate } = treeMateRef
-      switch (direction) {
-        case 'prev':
-          if (keyboardKey !== null) {
-            const node = treeMate.getPrev(keyboardKey, { loop: true })
-            if (node !== null) {
-              updateKeyboardKey(node.key)
-              updateHoverKey(node.key)
-              cascaderMenuInstRef.value?.scroll(
-                node.level,
-                node.index,
-                depx(optionHeightRef.value)
-              )
-            }
-          }
-          break
-        case 'next':
-          if (keyboardKey === null) {
-            const node = treeMate.getFirstAvailableNode()
-            if (node !== null) {
-              updateKeyboardKey(node.key)
-              updateHoverKey(node.key)
-              cascaderMenuInstRef.value?.scroll(
-                node.level,
-                node.index,
-                depx(optionHeightRef.value)
-              )
-            }
-          }
-          else {
-            const node = treeMate.getNext(keyboardKey, { loop: true })
-            if (node !== null) {
-              updateKeyboardKey(node.key)
-              updateHoverKey(node.key)
-              cascaderMenuInstRef.value?.scroll(
-                node.level,
-                node.index,
-                depx(optionHeightRef.value)
-              )
-            }
-          }
-          break
-        case 'child':
-          if (keyboardKey !== null) {
-            const currentNode = treeMate.getNode(keyboardKey)
-            if (currentNode !== null) {
-              if (currentNode.shallowLoaded) {
-                const node = treeMate.getChild(keyboardKey)
-                if (node !== null) {
-                  updateHoverKey(node.key)
-                  updateKeyboardKey(node.key)
-                }
-              }
-              else {
-                const { value: loadingKeySet } = loadingKeySetRef
-                if (!loadingKeySet.has(keyboardKey)) {
-                  addLoadingKey(keyboardKey)
-                  updateHoverKey(keyboardKey)
-                  const { onLoad } = props
-                  if (onLoad) {
-                    onLoad(currentNode.rawNode)
-                      .then(() => {
-                        deleteLoadingKey(keyboardKey)
-                      })
-                      .catch(() => {
-                        deleteLoadingKey(keyboardKey)
-                      })
-                  }
-                }
-              }
-            }
-          }
-          break
-        case 'parent':
-          if (keyboardKey !== null) {
-            const node = treeMate.getParent(keyboardKey)
-            if (node !== null) {
-              updateHoverKey(node.key)
-              updateKeyboardKey(node.key)
-            }
-          }
-          break
-      }
-    }
     function handleKeydown(e: KeyboardEvent): void {
       if (
         props.filterable
@@ -575,7 +405,7 @@ export default defineComponent({
                   doUncheck(keyboardKey)
                 }
                 else {
-                  const checkIsValid = doCheck(keyboardKey)
+                  const checkIsValid = doCheck(keyboardKey, focusSelectionInput)
                   if (!props.multiple && checkIsValid) {
                     closeMenu(true)
                   }
@@ -732,13 +562,6 @@ export default defineComponent({
         }
       }
     }
-    const showCheckboxRef = computed(() => {
-      if (props.multiple && props.cascade)
-        return true
-      if (mergedCheckStrategyRef.value !== 'child')
-        return true
-      return false
-    })
     provide(cascaderInjectionKey, {
       slots,
       mergedClsPrefixRef,
@@ -813,50 +636,6 @@ export default defineComponent({
         }
       }
     }
-    const cssVarsRef = computed(() => {
-      const {
-        self: {
-          optionArrowColor,
-          optionTextColor,
-          optionTextColorActive,
-          optionTextColorDisabled,
-          optionCheckMarkColor,
-          menuColor,
-          menuBoxShadow,
-          menuDividerColor,
-          menuBorderRadius,
-          menuHeight,
-          optionColorHover,
-          optionHeight,
-          optionFontSize,
-          loadingColor,
-          columnWidth
-        },
-        common: { cubicBezierEaseInOut }
-      } = themeRef.value
-      return {
-        '--n-bezier': cubicBezierEaseInOut,
-        '--n-menu-border-radius': menuBorderRadius,
-        '--n-menu-box-shadow': menuBoxShadow,
-        '--n-menu-height': menuHeight,
-        '--n-column-width': columnWidth,
-        '--n-menu-color': menuColor,
-        '--n-menu-divider-color': menuDividerColor,
-        '--n-option-height': optionHeight,
-        '--n-option-font-size': optionFontSize,
-        '--n-option-text-color': optionTextColor,
-        '--n-option-text-color-disabled': optionTextColorDisabled,
-        '--n-option-text-color-active': optionTextColorActive,
-        '--n-option-color-hover': optionColorHover,
-        '--n-option-check-mark-color': optionCheckMarkColor,
-        '--n-option-arrow-color': optionArrowColor,
-        '--n-menu-mask-color': changeColor(menuColor, { alpha: 0.75 }),
-        '--n-loading-color': loadingColor
-      }
-    })
-    const themeClassHandle = inlineThemeDisabled
-      ? useThemeClass('cascader', undefined, cssVarsRef, props)
-      : undefined
     return {
       ...exposedMethods,
       handleTriggerResize,
